Se confirma la configuracion del plazo de apertura al agregar eliminar o modificar item del plan de accion

+++ AppAc.Application/ItemPlanService.cs
@@ -11,23 +11,32 @@ namespace AppAc.Application
     {
         private readonly IUnitOfWork _unitOfWork;
         private readonly IPlanAccionRepository _planAccionRepository;
+        private readonly IPlazoAperturaRepository _plazoAperturaRepository;
         private readonly IItemPlanRepository _itemPlanRepository;
 
         public ItemPlanService(
            IUnitOfWork unitOfWork,
            IPlanAccionRepository planAccionRepository,
-           IItemPlanRepository itemPlanRepository
+           IItemPlanRepository itemPlanRepository,
+           IPlazoAperturaRepository plazoAperturaRepository
        )
         {
             _unitOfWork = unitOfWork;
             _itemPlanRepository = itemPlanRepository;
             _planAccionRepository = planAccionRepository;
+            _plazoAperturaRepository = plazoAperturaRepository;
         }
         public ItemPlanResponse RegistrarItem(ItemPlanRequest request)
         {
+
             var planAccion = _planAccionRepository.Find(request.PlanId);
             if (planAccion == null)
                 return new ItemPlanResponse("No se encontró el plan de acción",null);
+            var verificarPlazoAperturaService = new VerificarPlazoAperturaService(_plazoAperturaRepository);
+            var verificacionPlazoApertura = verificarPlazoAperturaService.Handle(planAccion.Actividad.Asignador.Identificacion);
+            if (verificacionPlazoApertura.Contains("Error"))
+                return new ItemPlanResponse(verificacionPlazoApertura, null);
+            
             var errors = ItemPlanUtils.CanConvertToItemPlan(request);
             if (errors.Any())
             {
@@ -52,9 +61,14 @@ namespace AppAc.Application
         public ItemPlanResponse EliminarItem(int id)
         {
             var item = _itemPlanRepository.Find(id);
+
             if (item == null)
                 return new ItemPlanResponse("No se encontró el item",item);
-            
+            var plan = _planAccionRepository.Find(item.PlanAccionId);
+            var verificarPlazoAperturaService = new VerificarPlazoAperturaService(_plazoAperturaRepository);
+            var verificacionPlazoApertura = verificarPlazoAperturaService.Handle(plan.Actividad.Asignador.Identificacion);
+            if (verificacionPlazoApertura.Contains("Error"))
+                return new ItemPlanResponse(verificacionPlazoApertura, null);
             _itemPlanRepository.Delete(item);
             _unitOfWork.Commit();
             return new ItemPlanResponse("Se elimino el item",item);
@@ -63,6 +77,13 @@ namespace AppAc.Application
         public ItemPlanResponse ModificarItem(ItemPlanUpdateRequest request)
         {
             var item = _itemPlanRepository.Find(request.Id);
+            if (item == null)
+                return new ItemPlanResponse("No se encontro el item", null);
+            var plan = _planAccionRepository.Find(item.PlanAccionId);
+            var verificarPlazoAperturaService = new VerificarPlazoAperturaService(_plazoAperturaRepository);
+            var verificacionPlazoApertura = verificarPlazoAperturaService.Handle(plan.Actividad.Asignador.Identificacion);
+            if (verificacionPlazoApertura.Contains("Error"))
+                return new ItemPlanResponse(verificacionPlazoApertura, null);
             var errors = new List<string>();
             errors.AddRange(ItemPlanUtils.CanConvertToItemPlan(request));
             if (errors.Any())
